Move CustomMessages into moderation
