+++ app/Discord/Moderation/Commands/AddLevelMessage.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace App\Discord\CustomMessages\Commands;
+namespace App\Discord\Moderation\Commands;
 
 use App\Discord\Core\Builders\EmbedFactory;
 use App\Discord\Core\Models\Guild;
 use App\Discord\Core\SlashCommand;
-use App\Discord\CustomMessages\Models\CustomMessage;
-use App\Discord\CustomMessages\Enums\CustomMessage as CustomMessageEnum;
+use App\Discord\Moderation\Enums\CustomMessage as CustomMessageEnum;
+use App\Discord\Moderation\Models\CustomMessage;
 use App\Discord\Roles\Enums\Permission;
 use Discord\Builders\MessageBuilder;
 use Discord\Parts\Interactions\Command\Option;
