+++ app/Discord/Moderation/Commands/DeleteLevelMessage.php
@@ -1,17 +1,17 @@
 <?php
 
-namespace App\Discord\CustomMessages\Commands;
+namespace App\Discord\Moderation\Commands;
 
 use App\Discord\Core\Builders\EmbedFactory;
 use App\Discord\Core\Models\Guild;
 use App\Discord\Core\SlashCommand;
-use App\Discord\CustomMessages\Models\CustomMessage;
+use App\Discord\Moderation\Models\CustomMessage;
 use App\Discord\Roles\Enums\Permission;
 use Discord\Builders\MessageBuilder;
 use Discord\Parts\Interactions\Command\Option;
 use Discord\Parts\Interactions\Interaction;
 
-class DeleteWelcomeMessage extends SlashCommand
+class DeleteLevelMessage extends SlashCommand
 {
     public function permission(): Permission
     {
@@ -25,11 +25,11 @@ class DeleteWelcomeMessage extends SlashCommand
 
     public function __construct()
     {
-        $this->description = __('bot.slash.delete-welcome-msg');
+        $this->description = __('bot.slash.delete-level-msg');
         $this->slashCommandOptions = [
             [
-                'name' => 'message',
-                'description' => __('bot.message'),
+                'name' => 'level',
+                'description' => __('bot.level'),
                 'type' => Option::INTEGER,
                 'required' => true,
                 'autocomplete' => true,
@@ -41,25 +41,25 @@ class DeleteWelcomeMessage extends SlashCommand
     public function action(): MessageBuilder
     {
         $customMessage = CustomMessage::where([
-            'id' => $this->getOption('message'),
+            'level' => $this->getOption('level'),
             'guild_id' => Guild::get($this->guildId)->id,
         ])->first();
 
         if (!$customMessage) {
-            return EmbedFactory::failedEmbed($this, __('bot.msg.welcome.not-found', ['message' => $this->getOption('message')]));
+            return EmbedFactory::failedEmbed($this, __('bot.msg.level.not-found', ['level' => $this->getOption('level')]));
         }
 
         $customMessage->delete();
-        return EmbedFactory::successEmbed($this, __('bot.msg.welcome.deleted'));
+        return EmbedFactory::successEmbed($this, __('bot.msg.level.deleted', ['level' => $this->getOption('level')]));
     }
 
     public function autoComplete(Interaction $interaction): array
     {
-        return CustomMessage::welcome($interaction->guild_id)->where('id', 'LIKE', "%{$this->getOption('message')}%")
+        return CustomMessage::level($interaction->guild_id)->where('level', 'LIKE', "%{$this->getOption('level')}%")
             ->limit(25)
             ->get()
             ->map(function ($modelInstance) {
-                return ['name' => substr($modelInstance->message, 0, 32), 'value' => $modelInstance->id];
+                return ['name' => substr($modelInstance->message, 0, 32), 'value' => $modelInstance->level];
             })->toArray();
     }
 }
