+++ app/Discord/Moderation/Commands/WelcomeMessagesIndex.php
@@ -1,17 +1,16 @@
 <?php
 
-namespace App\Discord\CustomMessages\Commands;
+namespace App\Discord\Moderation\Commands;
 
 use App\Discord\Core\Builders\EmbedBuilder;
 use App\Discord\Core\SlashIndexCommand;
-use App\Discord\CustomMessages\Enums\CustomMessage as CustomMessageEnum;
-use App\Discord\CustomMessages\Models\CustomMessage;
+use App\Discord\Moderation\Enums\CustomMessage as CustomMessageEnum;
+use App\Discord\Moderation\Models\CustomMessage;
 use App\Discord\Roles\Enums\Permission;
-use Discord\Builders\MessageBuilder;
 use Discord\Parts\Embed\Embed;
 use Discord\Parts\Interactions\Interaction;
 
-class LevelMessagesIndex extends SlashIndexCommand
+class WelcomeMessagesIndex extends SlashIndexCommand
 {
     public function permission(): Permission
     {
@@ -25,7 +24,7 @@ class LevelMessagesIndex extends SlashIndexCommand
 
     public function __construct()
     {
-        $this->description = __('bot.slash.level-index');
+        $this->description = __('bot.slash.welcome-index');
         parent::__construct();
     }
 
@@ -33,17 +32,17 @@ class LevelMessagesIndex extends SlashIndexCommand
     {
         $this->total = CustomMessage::byGuild($this->guildId)->count();
         $description = "";
-        foreach (CustomMessage::byGuild($this->guildId)->where('type', CustomMessageEnum::LEVEL->value)->skip($this->getOffset($this->getLastUser()))->limit($this->perPage)->get() as $message) {
-            $description .= "**{$message->level}** • {$message->message} \n";
+        foreach (CustomMessage::byGuild($this->guildId)->where('type', CustomMessageEnum::WELCOME->value)->skip($this->getOffset($this->getLastUser()))->limit($this->perPage)->get() as $message) {
+            $description .= "**{$message->id}** • {$message->message} \n";
         }
         return EmbedBuilder::create($this)
-            ->setTitle(__('bot.msg.level.title'))
+            ->setTitle(__('bot.msg.welcome.title'))
             ->setDescription($description)
             ->getEmbed();
     }
 
     public function autoComplete(Interaction $interaction): array
     {
-       return [];
+        return [];
     }
 }
